+++ src/components/mod.rs
@@ -2,3 +2,4 @@ pub mod central_panel;
 pub mod top_bar;
 pub mod json_viewer;
 pub mod theme;
+pub mod drag_and_drop;
